Split the components to their own files
They were also turned into ES6, in order to make use of the newest
features in the JavaScript world.
