+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["react"]
+  "presets": ["es2015", "react"]
 }
