+++ js/application.js
@@ -1,0 +1,5 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Game from './components/game'
+
+ReactDOM.render(<Game />, document.getElementById('game'))
