+++ package.json
@@ -10,6 +10,7 @@
   "devDependencies": {
     "babel-core": "^6.10.4",
     "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "eslint": "^3.0.1",
     "eslint-config-standard": "^5.3.1",
